requirements.txt
requests
pandas
numpy
streamlit
codecarbon
sentence_transformers
faiss-cpu